V-0.120

+++ Outil_Azur_complet/Theme/italk.cs
@@ -3101,7 +3101,7 @@ namespace iTalk
         }
 
         #endregion
-
+       
         public void AddRichTextBox()
         {
             var _RTB = iTalkRTB;
@@ -4409,8 +4409,8 @@ namespace iTalk
                     {
                         graphics.FillEllipse(brush2, 0x18, 0x18, (this.Width - 0x30) - 1, (this.Height - 0x30) - 1);
                     }
-                    SizeF MS = graphics.MeasureString(Convert.ToString(Convert.ToInt32((100 / _Maximum) * _Value)), Font);
-                    graphics.DrawString(Convert.ToString(Convert.ToInt32((100 / _Maximum) * _Value)), Font, Brushes.White, Convert.ToInt32(Width / 2 - MS.Width / 2), Convert.ToInt32(Height / 2 - MS.Height / 2));
+                    SizeF MS = graphics.MeasureString(Convert.ToString(Convert.ToInt64((100 / _Maximum) * _Value)), Font);
+                    graphics.DrawString(Convert.ToString(Convert.ToInt64((100 / _Maximum) * _Value)), Font, Brushes.White, Convert.ToInt64(Width / 2 - MS.Width / 2), Convert.ToInt64(Height / 2 - MS.Height / 2));
                     e.Graphics.DrawImage(bitmap, 0, 0);
                     graphics.Dispose();
                     bitmap.Dispose();
